+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,8 +1,8 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '6.3.1'.freeze
-    EXTENSION_BUILD = '202502120700'.freeze
+    EXTENSION_VERSION = '6.3.2-dev'.freeze
+    EXTENSION_BUILD = '202504150611'.freeze
 
     DEFAULT_LANGUAGE = 'en'
     ENABLED_LANGUAGES = %w[ar cs de en es fr he it nl pt ru uk zh vi]